type fixes

+++ Project1/GameObject.cpp
@@ -34,8 +34,8 @@ CGameObject::CGameObject(CPokeOrbitApp * pokeOrbit, wstring filename)
 	SetImage(filename);
 	//mersenne twister using time as a seed to decide what type of object to emit
 	mt19937_64 rand(chrono::system_clock::now().time_since_epoch().count());
-	mX = rand()%300 + 20;
-	mY = rand()%((int)sqrt(250000-mX*mX)-200) + 20;
+	mX = (double)(rand()%300 + 20);
+	mY = (double)(rand()%((int)sqrt(250000-mX*mX)-200) + 20);
 }
